- ajustando lógica no service produto

+++ src/main/java/br/edu/unicesumar/sistemaECommerce/controller/ProdutoController.java
@@ -27,24 +27,34 @@ public class ProdutoController {
     @FXML
   public void updateProduto() {
     Produto p = new Produto();
+    if( p != null){
     p.setId(1);
     p.setNome("");
     p.setPreco(15);
     p.setDescricao("");
 
     produtoService.updateProduto(p);
-
     label.setText("Produto atualizado com sucesso!");
 }
+    else{
+        label.setText("Produto não encontrado");
+}
+  }
 
 //Deletar
  @FXML
 public void deleteProduto() {
     Produto p = new Produto();
-    p.setId(1); 
+    if( p != null){
 
     produtoService.deleteProduto(p);
-
     label.setText("Produto removido com sucesso!");
+    }
+    else{
+        label.setText("Produto não encontrado");
+    }
+   
+
+
 }
 }
